Revert "progress implementing feature add substract logic"
