--- .github/workflows/firebase-hosting-pull-request.yml
@@ -1,26 +1,0 @@
-# This file was auto-generated by the Firebase CLI
-# https://github.com/firebase/firebase-tools
-
-name: Deploy to Firebase Hosting on PR
-on:
-  pull_request:
-jobs:
-  build_and_preview:
-    if: '${{ github.event.pull_request.head.repo.full_name == github.repository }}'
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v3
-      - name: Install Expo CLI
-        run: npm install -g expo-cli
-      - name: Install Babel dependencies
-        run: npm install --save-dev babel-preset-expo @babel/core @babel/preset-env
-      - name: Update Babel configuration
-        run: echo '{"presets": ["@babel/preset-env", "babel-preset-expo"]}' > .babelrc
-      - run: npx expo build:web
-      - uses: FirebaseExtended/action-hosting-deploy@v0
-        with:
-          repoToken: '${{ secrets.GITHUB_TOKEN }}'
-          firebaseServiceAccount: '${{ secrets.FIREBASE_SERVICE_ACCOUNT_CODELESS_APP_GENERATOR }}'
-          projectId: codeless-app-generator
-
-
