+++ package.json
@@ -11,13 +11,10 @@
   "dependencies": {
     "@expo/webpack-config": "^0.17.0",
     "@firebase/storage": "^0.9.13",
-    "@fortawesome/free-solid-svg-icons": "^6.4.0",
-    "@fortawesome/react-fontawesome": "^0.2.0",
     "@octokit/core": "^4.2.0",
     "@react-native-community/checkbox": "^0.5.14",
     "@react-native-community/google-signin": "^5.0.0",
     "axios": "^1.3.2",
-    "build": "^0.1.4",
     "cohere-ai": "^6.1.0",
     "expo": "~46.0.16",
     "expo-application": "~4.2.2",
@@ -35,12 +32,11 @@
     "picker": "^0.1.4",
     "react": "18.0.0",
     "react-dom": "18.0.0",
-    "react-native": "0.69.9",
+    "react-native": "0.69.6",
     "react-native-config": "^1.5.0",
     "react-native-google-signin": "^2.1.1",
     "react-native-select-multiple": "^2.1.0",
     "react-native-web": "~0.18.7",
-    "react-scripts": "^5.0.1",
     "yarn": "^1.22.19"
   },
   "devDependencies": {
